Do not allow removal of a voter to dip #voters below the min-required.
Summary: The voter distribution is used to compute minimum replication
factor/quorum for commit. When we remove voters, we need to check that we
are not going below the MIN-RF. So for 6 voters, MIN-RF=4 and we cannot
allow more than 2 voters to be removed from a healthy state of 6 voters.

We do this check for Flexi-Raft and in SINGLE_REGION_DYNAMIC mode of FR,
we narrow this check to the LEADER region.

Test Plan: Wrote a unit-test on mysql raft plugin side, which creates
a ring of 6 voters in LEADER region and tries to remove 3.

Reviewers: bhatvinay, yichenshen, yashtc

Subscribers:

Tasks:

Tags: