[kuduraft] Propagate candidate peer information
Summary:
When an election is started, the candidate populates its own peer_pb
into the new context that gets passed through RequestVote() rpc. This
enables better logging on other ring menebers. For now, the other peers
only log the host-port. Note that if the candidate is already in the active
config tracked by a peer, then the peer (responding to RequestVote())
already has the host-port. This change is more useful when the candidate
is not in the active config - makes debugging easier.

In future the same context can be used to exchange other information
like sql lag, system stats to make a more informed decision in election.

In a subsequent diff, I will make the change where the candidate does an
exponential backoff (after losing an election) before starting a new
one. This needs more thought since the protocol says that a peer should
respond 'yes' to a candidate who is not in its own active config, but is
caught up and has the latest log (basic necessity of config change
protocol).

Test Plan: build plugin and run tests

Reviewers: arahut

Subscribers:

Tasks:

Tags: