[wip] [kuduraft] aggressive backoffs in leader elections from a candidate not in config
Summary:
Background: When a node in the raft ring is removed from the config,
there is nothing in raft protocol that dictates that the removed node
should be notified. In fact, the protocol dictates that a config change
should be acted on immediately even before the config change itself is
committed - this essentially means that the config change message itself
is never sent to the node that is being removed. One should rely on
external entities to quiesce the removed node. This has been an area of
problem, where the removed node is not getting quiesced by external
entities for prolonged duration and ends up starting pre-election and
thus causing much noise in the system. Multiple fixes have gone in, but
this still keeps happening and makes debugging a little harder. Hence
this PR tries to address the situation by adding aggressive exponential
backoffs when a candidate loses a pre-election. This is a little tricky
because the protocol dictates that a voter should respond positively to
a candidate who is not in the voter's active-config, but still presents
the longest log (highest OpId).

VoteResponse is enhanced to have additional voter context. For now, the
only context is a boolen indicating if the candidate was removed from the
voter's committed config. When a voter responds to RequestVote() rpc (either
positively or negatively), the voter also sets this flag appropriately.

The candidate collects this response from every voter. If any of the
voter responds negatively and also indicates that the candidate was removed
from the voter's committed config, then the candidate keeps track of
this information.

If the candidate loses the election, and if atleast one voter indicated that
the candidate is removed from its committed config, then this
information is propagated to the upper layer. Based on this information, the
failure detector is snoozed more aggressively than in a regular ending of an
election (starting with 5 seconds, 25 secs, 125 secs and so on and ultimately
disabled)

Test Plan: build plugin with local tp2 and run tests

Reviewers: arahut

Subscribers:

Tasks:

Tags: