[kuduraft] [proxy] auto generate proxy topology using raft configs
Summary:
Manual updation of proxy topology (by external automation) needs proper
cordination across different systems as and when new peers are added,
removed or modified. Failure handling becomes tricky and difficult. So, it was
decided that it is better to  generate 'proxy topology' and routing
tables based on simple rules through raft configs. The routing rules
gets updated automatically when new config are proposed by the leader. This PR
tries to implement the same.

Different routing policies are defined in proxy_policy.h. There are two
policies that are defined and implemented:
1. Durable routing policy - existing policy where the routing topology
   is explicitly specified by external entities.
2. Simple region based routing policy - A new rule based proxy routing. It
   tries to keep parity with current vanilla deployment. At a high level,
   the rules used are:
 a) Each region has at most one designated 'proxy_peer' - it is chosen to be the
    peer which has a backing database. Leader ships messages to this
    'proxy_peer' when it has to send messages to any other peer in this
    region.
 b) Some regions may not have a 'proxy_peer'. Leader ships messages
    directly to all peers in such regions.
 c) There could be multiple 'candidate' proxy_peers in a given region - only
    one is chosen to be the real 'proxy_peer' for the region. This
    should not be required when region splitting is eventually
    implemented.
 d) When a new peer is added, we do not change any existing peer's proxy
    routing. This is to avoid churns in stable proxy routes on unrelated
    config changes

Implementation Notes:
1. An interface is defined through "IRoutingTable". The interface needs to be
implemented to support various routing policies.
2. A container holding all the roting policies is intantiated during
startup.
3. The policy can be shifted dynamically.

The container method was used to avoid additional locking overhead while
also supporting dynamic changes to proxy policy. The default policy is retained
to be 'DurableRoutingPolicy' to make it compatible with existing
implementation.

As the name suggests, DurableRoutingTable is persisted in cmeta - hence it is
continued to be instantiated and owned by cmeta_manager.
SimpleRegionRoutingTable is not persisted and is rebuilt using configs
on-the-fly. Hence it need not be managed by cmeta_manager.

Rules for SimpleRegionRoutingPolicy should get us a long way into
production deployment. In future, we could implement new policy with
rules for 'region-proying-another-region' to reduce cross region traffic (for
example cross atlantic traffic). If an external entity needs to implant (and
override) a topology, then the policy needs to be changed.
SimpleRegionRoutingPolicy is not persisted and hence it does not make sense to
allow external entities to provide/override arbitrary routing topology.

Test Plan: Tests in fbcode

Reviewers: arahut, mpercy

Subscribers:

Tasks:

Tags: