Add support for persisting variables
Summary:
When a configuration/variable change needs to be persisted, we will now write the variable/value to a new "persistent_vars" proto file. This file will be read during the `Init()` phase of consensus. Actions can be taken based on the values read from the file.

As an example, we add a new variable called `allow_start_election`. Whenever this variable is updates, we will write this to the proto file. On disabling the variable, the peer will not start any election - even if there are heartbeat failures from the leader or is an election is manually requested. Disabling this on all voters in the ring essentially becomes a way to block any elections from going through. Note that if this variable is disabled on the leader and then the leader is restarted, since it comes up as a follower, it won't get a chance to become leader again if for some reason there no other leader is elected in the ring.

Test Plan:
Tested as part of integration tests in fbcode

Reviewers: vinaybhat, arahut, abhinavsharma, mpercy, chili

Subscribers:

Tasks:

Tags: