[proxy] fix a crash in HandleProxyRequest
Summary:
"RaftConsensus::HandleProxyRequest()" builds a temporary vectors of
"ReplicateRefPtr" by reading ReplicateMsg from the log. It subsequently
builds the request to be forwarded to the destination peer. The ops in
the request are populated by calling
"downstream_request.mutable_ops()->AddAllocated(messages[i]->get()".
'downstream_request' is the request to be forwarded to the destination peer and
'messages' is the vector of "ReplicateMsg" built earlier. As a result of
this call, the ownership of 'ReplicateMsg' is passed over to the internal
protobuf implementation. However, when 'messages' goes out of scope, the
destructor could end up freeing the underlying memory for "ReplicateRefPtr".
This is dangerous - best case is an immediate crash due to a double free
and the worst case is silent memory corruption.

This PR fixes the problem by ensuring that vector 'messages' stays in
scope until the request completes. downstream_request->ExtractSubrange()
runs before "RaftConsensus::HandleProxyRequest()" exits, ensuring that
protobuf implementation does not free these messages.

Test Plan:
Wrote a fbcode stress test which repros the crash once in 2-3 runs. Ran
the test multiple times with the fix and saw no crash. Will check in the
test when I pull the changes into tp2.

Reviewers: arahut

Subscribers:

Tasks:

Tags: