[kuduraft] degrade a unhealthy peer to 'status-only' heartbeats
Summary:
In raft, the leader maintains persistent state about all nodes in the ring.
Whenever a ring goes down, leader will keep attempting to make connection to
this failed node as part of heartbeats (configured to be 500ms).

As part of sending a heartbeat, the leader tries to ship as many trxs as
possible (based on the batch size). The trxs to be shipped is determined based
on the leader's knowledge of the 'last-recieved-trx' by this particular failed
node. As more writes happen in the system, the failed node starts to 'lag' more.
Hence each successive attempt by the leader to connect to this particular
failed node will have to do more work - i.e read more trxs from the
log-cache/log (upto maximum configured batch size, defaulted to 2MB in MyRaft),
build the RPC message and try to push it to the failed node (which fails again,
because the node is not reachable). This is wasted work on the leader and on a
high write rate system this will soon start effecting leader's performance as
this involves opening a file, seeking to the desired op and reading 2MB
worth of data (in addition to building the request).

This PR tries to solve the problem by degrading the failed peer to a
'status-only' heartbeat request. What this essentially does is skipping
building the ops by reading from the log. Leader already tracks
'failed_attempts' i.e number of consecutive failure on communicating with a
peer. Based on this knowledge, leader determines that the peer should be
degraded to 'status-only' heartbeat request. When the leader is able to
communicate successfully with the peer again, 'failed_attempts' counter
is reset and the leader starts shipping ops again transparently to this
peer again.

Test Plan:
Tested in mysql mtr. Simulated a scenario where one of the peers failed
in a three ring setup. Ensured that the leader will stop shipping ops to
the failed peer and degrade to status-only heartbeat requests. The
leader is still able to commit ops because of the votes from the healthy
peer. When the failed peer recovers, leader transparently started
shipping ops again can caught up the recovered peer.

Reviewers: arahut

Subscribers:

Tasks:

Tags: