[kuduraft] [proxy] make proxy peer send heartbeats to the destination peer
Summary:
A proxy node might fail to build the proxy request for a destination
node due to multiple reasons. For example, the proxy node might not have
appended the ops needed to its own log yet. When this happens the
request should be degraded to heartbeat request i.e the proxy node should send
an empty heartbeat request to the destination node. However this was not
working as intended. The proxy peer would short circuit a failure and
return back to the leader. When this happens a few times, the
destination node's failure detector will trigger a pre-election round.
This has  a chance to destabilize the ring.

This PR resolves the issue by turning a failure to build request in the
proxy node into a heartbeat only request. This was the intended design
initially, but there was a bug which ended up short circuiting the
error and returned early.

Test Plan: Stress tests in fbcode by simulating network lag on proxy
peer

Reviewers: arahut

Subscribers:

Tasks:

Tags: