Skipping pending config checks when voter distribution adjustment is disabled
Summary: When voter distribution adjustment is disabled we're
essentially performing an unsafe-config change. In those scenarios we
should not crash or error out when there is a pending config change
present in the ring. These checks are already disabled for kudu's
inbuilt support for unsafe config change thru raft configuration.