[rpc-reactor] [perf] [2/n] ability to configure number of reactor threads
Summary:
This is needed when the number of peers in the ring is large. If the
value is specified in rpc options, then use it to override the value
specified through gflag.

Future steps:
1. Having more threads in learners/witness/secondaries is not really
useful. We could think of dynamically changing the number of reactor
threads as the ring goes through promotions and config changes.
2. Add support to create and destroy reactor threads during run time
(this will also help with 1)
3. Use different thread pools for reactor threads that serve peers which
are part of data commit quorum. This way thread priorities can be set
differently

Test Plan: tests in fbcode

Reviewers: arahut

Subscribers:

Tasks:

Tags: