Adding a function to dump the current certificates into a string vector
Summary: Also modifying LoadCertFiles to take in a parameter
for use_new_store, which is typically set to true for outside.

Test Plan: Incorporated these changes into a Mysql raft RPM
and used the DumpCertsInfo and LoadCertFiles on a timer thread
to refresh certificates from disk. New certicates are expected to
be planted through some external automation.

Reviewers: yashtc

Subscribers:

Tasks:

Tags: