Make transfer context non-optional
Summary:

Since we're using the context to store downtimes, we'll need it in both
the case where a LBU starts the promotion, as well as a typical LMP
where an external party starts it.

As such, we'll always have a ElectionContext, so clean this up to make
it compulsory and simpler.

Test Plan:

Make sure both kudu and fbcode plugin can build.

Reviewers: anirban-fb, bhatvinay, li-chi

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>