Adding proper locking for DumpCertFields
Summary: We saw a crash in raft at the time that certs are refreshed.
We were not taking a lock during the DumpCertFields call from outside,
which meant that the X509 which was gotten could have been released.

Fixed this by making sure that the entry point takes the tls_context
lock.

Test Plan: Tested on mysql raft side by writing a test which refreshes
certs continuously while reading them concurrently.

Reviewers: yichenshen

Subscribers:

Tasks:

Tags: