Add functions to define values of SO_SNDBUF and SO_RCVBUF
Summary:

Add a way to set SO_SNDBUF and SO_RCVBUF when the messenger starts a new
socket.

This allows us to configure the TCP buffer sizes.

Test Plan:

Start up a Raft ring with options configured and inspect the socket buf
sizes with `ss -tmp`. Values shown should be twice of what we set here.

Also run an election to make sure the values are still correct.

Currently there's no way to reset these values on the fly, but running
an election forces reshuffling of sockets, so we can change the value,
run an election, then check with `ss -tmp` again.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>