Add a handle in the messenger that can reset all connections
Summary:

This is intended to be called manually via an external API. It lets us
reset all connections so the any new connection/socket settings can be
applied when they're created again.

Test Plan:

Build in fbcode with a handle to trigger this, upload to staging, and
run it on both followers and leaders. Give it a while then check the
socket FDs to make sure they're changed. There shouldn't be any
disruptions to the ring other than that.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>