Add a raft_ring_token to persistent_vars
Summary:

This is setup as a lockless read var. Writes will need external
synchronization as seen in RaftConsensus.

I also added a server option to provide a initial token var. This
initial token will be set into persistent_vars if a token isn't already
present and can be used for boostrapping in an idempontent fashion (i.e.
if Raft is already bootstrapped and the token changed, we don't have to
worry about cleaning up the bootstrap settings).

Test Plan:

This will be used later. For now we just compile.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>