Add logic for UUID enforcement
Summary:

We'll pull out the local token out of Raft consensus for the three
internal RPCs. Then we'll check if the RPC's token matches up.

FLAGs_enforce_raft_rpc_token controls strict enforcement. If true, we'll
terminate the request before it ever reaches RaftConsensus should there
be a mismtach. If we're not enforcing, we log a WARNING and ignore the
mismatch

Test Plan:

Sync this to fbcode and implement APIs to set the token. We want to
ensure the ring responds the right way based on what we set.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>