Enforce that responses have a matching RPC token
Summary:

As mentioned in the previous commit, we need to check if the response
also has the correct token.

We do this by setting a flag in RPC controller. Then in peer proxy we
hijack the response with an error if the response's token doesn't match
the request's token.

Test Plan:

We can test this in fbcode. The motivating issue is that an instance
with a wrong token and is not enforcing a token match could cause a term
bump in other instances should it unilaterally start an election (the
election will fail because it's vote requests will be rejected, but it's
higher term will cause the current leader to step down).

Once we enforce matches for responses, then this shouldn't happen
anymore.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>