Withhold votes from hugely lagging voters in flexiraft SRD mode
Summary:

New Non Standard Voting Heuristic: In Flexi Raft, since quorum sizes are typically
small, we have much lesser tolerance for failed nodes. A partially failed
node can be a node which is lagging (say) 10 million opids behind the CANDIDATE.
In normal raft, it will give votes to the CANDIDATE, but if the VOTER is
in same region as CANDIDATE, the VOTER will also soon be part of the write
quorum for the No-Op and needs to be caught up. Hence giving VOTES eagerly
when SELF is much much behind the CANDIDATE, will not help No-Op commit.
So we add another heuristic to address this. Here is how it works
1. It withholds votes if SELF is much behind CANDIDATE.
2. Only applies to flexi raft Single Region dynamic mode
3. Heuristic has a kill switch lag_threshold_for_request_vote=-1.  (default)
4. Only applies to VOTERs which are in the same region as CANDIDATE

Test Plan: Create test in fbcode.

Reviewers:

Subscribers:

Tasks:

Tags: